requirements.txt
streamlit~=1.43.2
pandas~=2.2.3
ckanapi~=4.8
dotenv~=0.9.9
python-dotenv~=1.0.1
openpyxl~=3.1.5
altair~=5.5.0
google-analytics-data~=0.18.17
google-generativeai~=0.8.4
reportlab~=4.3.1